backend/requirements.txt
flask
flask-cors
google-genai
protobuf
requests
beautifulsoup4
python-dotenv
